+++ src/datasources/results/types/QueryResults.types.ts
@@ -6,7 +6,10 @@ export interface QueryResults extends ResultsQuery {
   useTimeRange?: boolean;
   recordCount?: number;
   queryBy?: string;
-  partNumberQuery?: string[];
+}
+
+export interface QueryResultsDefaultValues extends QueryResults {
+  useTimeRangeFor: string;
 }
 
 export interface QueryResultsDefaultValues extends QueryResults {
@@ -23,7 +26,6 @@ export interface ResultsVariableQuery extends ResultsQuery {
 }
 
 export interface StepsVariableQuery extends ResultsQuery {
-  partNumberQueryInSteps: string[];
   queryByResults: string;
   queryBySteps?: string;
   stepsTake?: number;