+++ src/datasources/results/types/QuerySteps.types.ts
@@ -4,10 +4,12 @@ export interface QuerySteps extends ResultsQuery {
   outputType: OutputType;
   properties?: StepsProperties[];
   partNumberQuery: string[];
+  orderBy?: string;
+  descending?: boolean;
   useTimeRange?: boolean;
   recordCount?: number;
   showMeasurements?: boolean;
-  resultsQuery?: string;
+  resultsQuery: string;
   stepsQuery?: string;
 }