+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+## [3.132.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.131.0...v3.132.0) (2025-06-12)
+
+### Features
+
+* **results:** revert Partnumber field from Steps Query ([#283](https://github.com/ni/systemlink-grafana-plugins/issues/283)) ([efbf378](https://github.com/ni/systemlink-grafana-plugins/commit/efbf3789180ef30969f946472403b88ca5d79e89))
+
+## [3.131.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.130.0...v3.131.0) (2025-06-12)
+
+### Features
+
+* **results:** revert partNumber field for results query ([#280](https://github.com/ni/systemlink-grafana-plugins/issues/280)) ([fa2638d](https://github.com/ni/systemlink-grafana-plugins/commit/fa2638d7df83092255b71a27201e53458e85a220))
+
+### Bug Fixes
+
+* **results:** update default limit of take to 1000 ([#282](https://github.com/ni/systemlink-grafana-plugins/issues/282)) ([404b630](https://github.com/ni/systemlink-grafana-plugins/commit/404b6309b50898cc4f251ea771f0008e84561ccf))
+
+## [3.130.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.129.1...v3.130.0) (2025-06-12)
+
+### Features
+
+* **results:** update measurements to be listed as separate columns ([#281](https://github.com/ni/systemlink-grafana-plugins/issues/281)) ([3da8a68](https://github.com/ni/systemlink-grafana-plugins/commit/3da8a68db244ab8890f775066ea97f18f88f11bb))
+
+## [3.129.1](https://github.com/ni/systemlink-grafana-plugins/compare/v3.129.0...v3.129.1) (2025-06-12)
+
+### Bug Fixes
+
+* **results:** outputs are not shown as query name ([#275](https://github.com/ni/systemlink-grafana-plugins/issues/275)) ([9925329](https://github.com/ni/systemlink-grafana-plugins/commit/9925329e1079949510648d64615d84f09b2dace7))
+
+## [3.129.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.128.1...v3.129.0) (2025-06-12)
+
+### Features
+
+* **core:** Add support for custom date time operations ([#267](https://github.com/ni/systemlink-grafana-plugins/issues/267)) ([0873fb2](https://github.com/ni/systemlink-grafana-plugins/commit/0873fb2dd98dab31a72b1c5b208a894ecece05b4))
+
+## [3.128.1](https://github.com/ni/systemlink-grafana-plugins/compare/v3.128.0...v3.128.1) (2025-06-12)
+
+### Bug Fixes
+
+* **workorders:** multiple API calls when editing panels ([#277](https://github.com/ni/systemlink-grafana-plugins/issues/277)) ([b6c2afe](https://github.com/ni/systemlink-grafana-plugins/commit/b6c2afe1813b7e113f1eb9c6b964e7bcddf93b05))
+
+## [3.128.0](https://github.com/ni/systemlink-grafana-plugins/compare/v3.127.4...v3.128.0) (2025-06-12)
+
+### Features
+
+* **results:** Add "has children" property in steps query builder ([#278](https://github.com/ni/systemlink-grafana-plugins/issues/278)) ([beefaee](https://github.com/ni/systemlink-grafana-plugins/commit/beefaee162d5bdd4a1d0f1eb863d46b9b2904037))
+
+## [3.127.4](https://github.com/ni/systemlink-grafana-plugins/compare/v3.127.3...v3.127.4) (2025-06-12)
+
+### Bug Fixes
+
+* **results,products:** multiple API calls when editing panels ([#274](https://github.com/ni/systemlink-grafana-plugins/issues/274)) ([64b1145](https://github.com/ni/systemlink-grafana-plugins/commit/64b1145cf03b53536aed2b42036c08255a0c76e3))
+
+## [3.127.3](https://github.com/ni/systemlink-grafana-plugins/compare/v3.127.2...v3.127.3) (2025-06-11)
+
+### Bug Fixes
+
+* **results:** Set "STARTED_AT" as the default option for the order by field ([#270](https://github.com/ni/systemlink-grafana-plugins/issues/270)) ([291f74e](https://github.com/ni/systemlink-grafana-plugins/commit/291f74e73d510c5a8d90d8ec87868e3dfea3baca))
+
 ## [3.127.2](https://github.com/ni/systemlink-grafana-plugins/compare/v3.127.1...v3.127.2) (2025-06-11)
 
 ### Bug Fixes